feat: handle iframes and init code when choosing elements
